backend/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-environ==0.11.2
django-filter==24.3
django-rest-knox==5.0.2
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
sqlparse==0.5.1
tzdata==2024.2
requests==2.32.3
pytz==2024.2
